Исправил все указанные ошибки
